何:修改了和auth id有关的问题，修改了ddl

+++ Documents/DDL_Proj2_1.1.sql
@@ -45,7 +45,7 @@ create table videos
     owner_mid   bigint    not null,   -- 视频所有者的用户ID
     commit_time timestamp not null,   -- 视频提交时间
     public_time timestamp,            -- 视频对外公开时间
-    duration    integer   not null,   -- 视频持续时间（秒）
+    duration    float   not null,   -- 视频持续时间（秒）
     description text,                 -- 视频简短介绍
     check ( bv like 'BV%')            -- 确保 'bv' 字段以 'BV' 开头
 );
